Sample project for testing
Since all of my django projects usually end up with a very similar
structure, I thought it would be useful to include some common settings in
this repository. This is useful for testig the cms app, but also as a
scaffold for new projects. The syntax in settings.py is intentionally
invalid, so you will have to make your desired changes before you can
actually use it.

My personal intended usecase: use this repository as a template for new
projects, edit settings.py and add custom apps.